# Conflicts:
#	.idea/workspace.xml
#	out/production/MonkeyPilot/MonkeyPilot$1.class
#	out/production/MonkeyPilot/MonkeyPilot$2.class
#	out/production/MonkeyPilot/MonkeyPilot.class
#	out/production/MonkeyPilot/controller/Planecontroller.class
#	src/MonkeyPilot.java
#	src/controller/Planecontroller.java
#	src/model/Model.java

+++ src/view/View.java
@@ -15,6 +15,6 @@ public class View {
         this.image = image;
     }
     public void draw(Graphics g, Model model){
-        g.drawImage(image,model.getX(),model.getY(),model.getWidth(),model.getHeight(),null);
+        g.drawImage(image,(int)model.getX(),(int)model.getY(),model.getWidth(),model.getHeight(),null);
     }
 }
